# docs/openapi-gpt-3.0.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "esdata API",
    "description": "API de consulta fiscal-regulatoria con trazabilidad a fuente oficial, superficies de governance y workers observables via sync_log.",
    "version": "0.1.6"
  },
  "servers": [
    {
      "url": "https://api.desuscribir.es",
      "description": "Production API"
    }
  ],
  "paths": {
...
    "/v1/legislacion": {
      "get": {
        "tags": [
          "legislacion"
        ],
        "summary": "List Legislacion",
        "operationId": "list_legislacion",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Tamano de pagina aplicado",
              "default": 200,
              "title": "Limit"
            },
            "description": "Tamano de pagina aplicado"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset de resultados",
              "default": 0,
              "title": "Offset"
            },
            "description": "Offset de resultados"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/legislacion/{codigo}": {
      "get": {
        "tags": [
          "legislacion"
        ],
        "summary": "Get law details",
        "operationId": "get_norma",
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Codigo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "description": "Get metadata for a Spanish law by code (e.g. LIVA, LIRPF, LIS). Returns title, jurisdiction, document type, and coverage status."
      }
    },
    "/v1/legislacion/{codigo}/articulos": {
      "get": {
        "tags": [
          "legislacion"
        ],
        "summary": "List Articulos",
        "operationId": "list_articulos",
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Codigo"
            }
          },
          {
            "name": "tipo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Tipo"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Tamano de pagina aplicado",
              "default": 200,
              "title": "Limit"
            },
            "description": "Tamano de pagina aplicado"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset de resultados",
              "default": 0,
              "title": "Offset"
            },
            "description": "Offset de resultados"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticulosListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/legislacion/{codigo}/articulos/{numero}": {
      "get": {
        "tags": [
          "legislacion"
        ],
        "summary": "Get article text",
        "operationId": "get_articulo",
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Codigo"
            }
          },
          {
            "name": "numero",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Numero"
            }
          },
          {
            "name": "vigente_en",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Vigente En"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "description": "Get the full text of a specific article from a law. Optionally filter by vigencia_en (YYYY-MM-DD) for historical versions. Returns the article text, validity dates, and confidence info."
      }
    },
    "/v1/legislacion/{codigo}/articulos/{numero}/historial": {
      "get": {
        "tags": [
          "legislacion"
        ],
        "summary": "Get Articulo Historial",
        "operationId": "get_articulo_historial",
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Codigo"
            }
          },
          {
            "name": "numero",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Numero"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "schemas": {
...
      "ArticuloListItem": {
        "properties": {
          "numero": {
            "type": "string",
            "title": "Numero",
            "description": "Número del artículo"
          },
          "titulo": {
            "type": "string",
            "title": "Titulo",
            "description": "Título del artículo"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo",
            "description": "Tipo (articulo, disposicion, etc.)"
          }
        },
        "type": "object",
        "required": [
          "numero",
          "tipo"
        ],
        "title": "ArticuloListItem"
      },
...
      "ArticulosListResponse": {
        "properties": {
          "norma": {
            "type": "string",
            "title": "Norma",
            "description": "Código de la norma"
          },
          "articulos": {
            "items": {
              "$ref": "#/components/schemas/ArticuloListItem"
            },
            "type": "array",
            "title": "Articulos"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Límite aplicado"
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "description": "Offset aplicado"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "Si existen más resultados"
          },
          "next_offset": {
            "type": "integer",
            "title": "Next Offset",
            "description": "Offset de continuación"
          }
        },
        "type": "object",
        "required": [
          "norma",
          "articulos"
        ],
        "title": "ArticulosListResponse"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Dedicated ESData API key for GPT Actions."
      }
    }
  }
}